Add initial support for custom UUID and badge services
A new UUID has been generated and added so that we can advertise
our own custom services. Right now there's just support for an unlocks
characteristic and a password to enable it. You must supply the correct
password after connecting via bluetooth in order to modify the unlocks
value. Once changed, the value is saved to the user config in flash.

Note: we must use an event handler to write out the change using the
main thread. We can't do it in the SoftDevice event dispatch thread
because we must respond to the peer device quickly when it sends a
write request, and doing a flash erase and update takes too long to
do it synchronously.

An initial GATTC module has also been added but it doesn't do anything
yet except provide a dummy dispatch routine. I needed a custom service
first before I could work on a client for it. More to come this weekend.

+++ software/firmware/badge_840/userconfig.h
@@ -63,7 +63,7 @@ typedef struct userconfig {
 
   uint16_t lastdeath; // last time you died
   uint8_t in_combat;
-  uint16_t unlocks;
+  uint32_t unlocks;
 
   int16_t hp;
   uint16_t xp;
